requirements.txt
aiogram>=3.3.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
PyYAML>=6.0
tzdata>=2024.1